Add max-product linear solution
I did not bother write the brute force solution this time.

The approach is similar to linear solution of max-sub-array, but with a
second product computed (alt product) to handle sign inversion, a
behavior that sum doesn't display. Also zero nullifying is unique to
this problem.

A version of this product problem that would have the same solution as
the sum problem would be if array contained only positive floats.