Improve 3sum performance by 1.67×
The dict with counts is replaced with a set, and the iteration to build
it has been removed. It's built during search loop, and counting is
useless because the values of the couple for which we seek the opposite
value are added to the set afterward. I.e. leverage the iteration order.

Also the triple is sorted in-place, which makes code run faster:
probably because less copying is needed, and less short-lived objects
are created (less stress on GC).