Add number-of-islands solution
The algorithm counts the number of times a new traversal has to be
started, this gives the number of islands. Islands are actually
connected component of the graph having 1 cells as vertices.

The complexity is O(V), i.e. O(m*n). Space complexity may be reduced to
constant (auxiliary) by using the input matrix itself to mark visited
nodes. But this would be a bit silly.