Add longest repeating char replacement linear solution
It uses a sliding window and maintains counts for each char value
contained in it. It grows the window size as long as the diff between
the window size and the ocurrences of the most common char is not more
than the allowed number of replacements.

It always advances the window end, which means it doesn't consider
windows shorter than the longest valid one it already found.