Improve group anagram solution to linear
Replace the sorting with a count array. This works fine because the
alaphabet size is known, thus constant. Using an array an not a hashmap
avoids the need to sort; the array is always in alphabet order
regardless of the order of the chars in the strings.

Using an ordered dict initialized with the alphabet chars always in the
same order would be a acceptable alternative. It would display the same
time and space complexities, despite most likely being slower.