Improve longest palindromic substring to quadratic time
Use memoization (dynamic programming) to ensure palindrome check is not
done more than once for each (i, j) pair, that is, not more that n²
times overall. The two nested loops are O(n²) and the plalindrome check
are O(n²), so overall O(2n²).

Space complexity is O(n²) for the memoization table.