lc: Add linear solution to next-greater-element
The binary search was actually detrimental. Treating the list of
potential next greater elements as a stack allows leveraging the
accessing and removal pattern to make the whole management of the list
linear: its elements are gobally accessed 2n times. This is the same
order as the preliminary and main loops, thus the time complexity of the
whole algorithm is O(n).

The space complexity stays the same at O(n).